backend/requirements.txt
Django==6.0.3

djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1

requests==2.32.5

# Queue/scheduler
celery>=5.3,<6
redis>=5,<6

# Observability
sentry-sdk>=2,<3
